+++ test/OneImlx.Terminal.Tests/Hosting/TerminalHostedServiceTests.cs
@@ -5,9 +5,16 @@
     https://terms.perpetualintelligence.com/articles/intro.html
 */
 
-using FluentAssertions;
+using System;
+using System.Linq;
+using System.Reflection;
+using System.Security.AccessControl;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using FluentAssertions;
 using OneImlx.Shared.Licensing;
 using OneImlx.Terminal.Commands;
 using OneImlx.Terminal.Commands.Checkers;
@@ -18,12 +25,6 @@ using OneImlx.Terminal.Licensing;
 using OneImlx.Terminal.Mocks;
 using OneImlx.Terminal.Runtime;
 using OneImlx.Terminal.Stores;
-using System;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 using Xunit;
 
 namespace OneImlx.Terminal.Hosting
@@ -76,34 +77,12 @@ namespace OneImlx.Terminal.Hosting
         }
 
         [Fact]
-        public void StartAsync_Default_ShouldPrint_LicenseInfo()
-        {
-            // use reflection to call
-            MethodInfo? printLic = defaultCliHostedService.GetType().GetMethod("PrintHostApplicationLicensingAsync", BindingFlags.Instance | BindingFlags.NonPublic);
-            Microsoft.VisualStudio.TestTools.UnitTesting.Assert.IsNotNull(printLic);
-            printLic.Invoke(defaultCliHostedService, [MockLicenses.TestLicense]);
-
-            logger.Messages.Should().BeEmpty();
-
-            mockTerminalConsole.Messages.Should().HaveCount(9);
-            mockTerminalConsole.Messages[0].Should().Be("tenant=test_name (test_tenantid)");
-            mockTerminalConsole.Messages[1].Should().Be("country=test_country");
-            mockTerminalConsole.Messages[2].Should().Be("license=test_id");
-            mockTerminalConsole.Messages[3].Should().Be("mode=test_mode");
-            mockTerminalConsole.Messages[4].Should().Be("deployment=test_deployment");
-            mockTerminalConsole.Messages[5].Should().Be("usage=urn:oneimlx:lic:usage:rnd");
-            mockTerminalConsole.Messages[6].Should().Be("plan=urn:oneimlx:terminal:plan:demo");
-            mockTerminalConsole.Messages[7].Should().StartWith("iat=");
-            mockTerminalConsole.Messages[8].Should().StartWith("exp=");
-        }
-
-        [Fact]
         public void StartAsync_Default_ShouldPrint_MandatoryLicenseInfo_For_Custom_RND()
         {
             License community = new(TerminalLicensePlans.Custom, LicenseUsage.RnD, "testkey", MockLicenses.TestClaims, MockLicenses.TestQuota);
 
             // use reflection to call
-            MethodInfo? printLic = defaultCliHostedService.GetType().GetMethod("PrintHostApplicationMandatoryLicensingAsync", BindingFlags.Instance | BindingFlags.NonPublic);
+            MethodInfo? printLic = defaultCliHostedService.GetType().BaseType!.GetMethod("PrintWarningIfDemoAsync", BindingFlags.Instance | BindingFlags.NonPublic);
             Microsoft.VisualStudio.TestTools.UnitTesting.Assert.IsNotNull(printLic);
             printLic.Invoke(defaultCliHostedService, [community]);
 
@@ -116,7 +95,7 @@ namespace OneImlx.Terminal.Hosting
             License community = new(TerminalLicensePlans.Demo, LicenseUsage.Educational, "testkey", MockLicenses.TestClaims, MockLicenses.TestQuota);
 
             // use reflection to call
-            MethodInfo? printLic = defaultCliHostedService.GetType().GetMethod("PrintHostApplicationMandatoryLicensingAsync", BindingFlags.Instance | BindingFlags.NonPublic);
+            MethodInfo? printLic = defaultCliHostedService.GetType().BaseType!.GetMethod("PrintWarningIfDemoAsync", BindingFlags.Instance | BindingFlags.NonPublic);
             Microsoft.VisualStudio.TestTools.UnitTesting.Assert.IsNotNull(printLic);
             printLic.Invoke(defaultCliHostedService, [community]);
 
@@ -132,7 +111,7 @@ namespace OneImlx.Terminal.Hosting
             License community = new(TerminalLicensePlans.Demo, LicenseUsage.RnD, "testkey", MockLicenses.TestClaims, MockLicenses.TestQuota);
 
             // use reflection to call
-            MethodInfo? printLic = defaultCliHostedService.GetType().GetMethod("PrintHostApplicationMandatoryLicensingAsync", BindingFlags.Instance | BindingFlags.NonPublic);
+            MethodInfo? printLic = defaultCliHostedService.GetType().BaseType!.GetMethod("PrintWarningIfDemoAsync", BindingFlags.Instance | BindingFlags.NonPublic);
             Microsoft.VisualStudio.TestTools.UnitTesting.Assert.IsNotNull(printLic);
             printLic.Invoke(defaultCliHostedService, [community]);
 
@@ -143,97 +122,12 @@ namespace OneImlx.Terminal.Hosting
         }
 
         [Fact]
-        public void StartAsync_Default_ShouldPrint_OnStarted()
-        {
-            // use reflection to call
-            MethodInfo? print = defaultCliHostedService.GetType().GetMethod("OnStarted", BindingFlags.Instance | BindingFlags.NonPublic);
-            Microsoft.VisualStudio.TestTools.UnitTesting.Assert.IsNotNull(print);
-            print.Invoke(defaultCliHostedService, null);
-
-            logger.Messages.Should().BeEmpty();
-
-            mockTerminalConsole.Messages.Should().HaveCount(1);
-            mockTerminalConsole.Messages[0].Should().StartWith("Application started on");
-        }
-
-        [Fact]
-        public void StartAsync_Default_ShouldPrint_OnStopped()
-        {
-            // use reflection to call
-            MethodInfo? print = defaultCliHostedService.GetType().GetMethod("OnStopped", BindingFlags.Instance | BindingFlags.NonPublic);
-            Microsoft.VisualStudio.TestTools.UnitTesting.Assert.IsNotNull(print);
-            print.Invoke(defaultCliHostedService, null);
-
-            logger.Messages.Should().BeEmpty();
-
-            mockTerminalConsole.Messages.Should().HaveCount(1);
-            mockTerminalConsole.Messages[0].Should().StartWith("Application stopped on");
-        }
-
-        [Fact]
-        public void StartAsync_Default_ShouldPrint_OnStopping()
-        {
-            // use reflection to call
-            MethodInfo? print = defaultCliHostedService.GetType().GetMethod("OnStopping", BindingFlags.Instance | BindingFlags.NonPublic);
-            Microsoft.VisualStudio.TestTools.UnitTesting.Assert.IsNotNull(print);
-            print.Invoke(defaultCliHostedService, null);
-
-            logger.Messages.Should().BeEmpty();
-
-            mockTerminalConsole.Messages.Should().HaveCount(1);
-            mockTerminalConsole.Messages[0].Should().Be("Stopping application...");
-        }
-
-        [Fact]
-        public async Task StartAsync_ShouldCallCustomizationInCorrectOrderAsync()
+        public async Task StartAsync_ShouldCallCustomizationAsync()
         {
             MockTerminalHostedServiceStaticCounter.Restart();
             await mockCustomCliHostedService.StartAsync(cancellationToken);
 
-            // #1 call
-            mockCustomCliHostedService.RegisterEventsCalled.Should().NotBeNull();
-            mockCustomCliHostedService.RegisterEventsCalled.Item1.Should().Be(1);
-            mockCustomCliHostedService.RegisterEventsCalled.Item2.Should().BeTrue();
-
-            // #2 call
-            mockCustomCliHostedService.PrintHostAppHeaderCalled.Should().NotBeNull();
-            mockCustomCliHostedService.PrintHostAppHeaderCalled.Item1.Should().Be(2);
-            mockCustomCliHostedService.PrintHostAppHeaderCalled.Item2.Should().BeTrue();
-
-            // #3 call
-            mockLicenseExtractor.ExtractLicenseCalled.Should().NotBeNull();
-            mockLicenseExtractor.ExtractLicenseCalled.Item1.Should().Be(3);
-            mockLicenseExtractor.ExtractLicenseCalled.Item2.Should().BeTrue();
-
-            // #4 call
-            mockCustomCliHostedService.PrintHostLicCalled.Should().NotBeNull();
-            mockCustomCliHostedService.PrintHostLicCalled.Item1.Should().Be(4);
-            mockCustomCliHostedService.PrintHostLicCalled.Item2.Should().BeTrue();
-
-            // #5 call
-            mockLicenseChecker.CheckLicenseCalled.Should().NotBeNull();
-            mockLicenseChecker.CheckLicenseCalled.Item1.Should().Be(5);
-            mockLicenseChecker.CheckLicenseCalled.Item2.Should().BeTrue();
-
-            // #6 call
-            mockCustomCliHostedService.PrintMandatoryLicCalled.Should().NotBeNull();
-            mockCustomCliHostedService.PrintMandatoryLicCalled.Item1.Should().Be(6);
-            mockCustomCliHostedService.PrintMandatoryLicCalled.Item2.Should().BeTrue();
-
-            // #7 call
-            mockOptionsChecker.CheckOptionsCalled.Should().NotBeNull();
-            mockOptionsChecker.CheckOptionsCalled.Item1.Should().Be(7);
-            mockOptionsChecker.CheckOptionsCalled.Item2.Should().BeTrue();
-
-            // #8 call
-            mockCustomCliHostedService.CheckAppConfigCalled.Should().NotBeNull();
-            mockCustomCliHostedService.CheckAppConfigCalled.Item1.Should().Be(8);
-            mockCustomCliHostedService.CheckAppConfigCalled.Item2.Should().BeTrue();
-
-            // #9 call
-            mockCustomCliHostedService.RegisterHelpArgumentCalled.Should().NotBeNull();
-            mockCustomCliHostedService.RegisterHelpArgumentCalled.Item1.Should().Be(9);
-            mockCustomCliHostedService.RegisterHelpArgumentCalled.Item2.Should().BeTrue();
+            mockCustomCliHostedService.ConfigureLifetimeCalled.Should().BeTrue();
         }
 
         [Fact]
@@ -294,32 +188,6 @@ namespace OneImlx.Terminal.Hosting
                 foundHelp.Should().BeFalse();
                 helpAttr.Should().BeNull();
             }
-        }
-
-        [Fact]
-        public async Task StartAsync_ShouldRegister_Application_EventsAsync()
-        {
-            IHostApplicationLifetime hostApplicationLifetime = host.Services.GetRequiredService<IHostApplicationLifetime>();
-
-            // Create a host builder with mock event hosted service
-            hostBuilder = Host.CreateDefaultBuilder();
-            hostBuilder.ConfigureServices(services =>
-            {
-                // Make sure we use the instance created for test
-                services.AddHostedService(EventsHostedService);
-            });
-
-            mockCliEventsHostedService.OnStartedCalled.Should().BeFalse();
-            host = await hostBuilder.StartAsync();
-            mockCliEventsHostedService.OnStartedCalled.Should().BeTrue();
-
-            mockCliEventsHostedService.OnStoppingCalled.Should().BeFalse();
-            mockCliEventsHostedService.OnStoppedCalled.Should().BeFalse();
-            hostApplicationLifetime.StopApplication();
-            mockCliEventsHostedService.OnStoppingCalled.Should().BeTrue();
-
-            // TODO: OnStopped not called, check with dotnet team
-            //mockCliEventsHostedService.OnStoppedCalled.Should().BeTrue();
         }
 
         [Fact]